// src/utils/types.ts
export interface Clause<T extends Entity> {
  /** The entity field to query or update */
  field: keyof T;
  /** * The value used in the clause.
   * It is either the property type, a raw value for IN ([number[]]), or null.
   */
  value: T[keyof T] | any[] | null; // Use `any[]` for 'IN' array values
}

/**
 * Supported operators for WHERE clause conditions.
 */
type WhereOperator =
  | "="
  | "!="
  | ">"
  | "<"
  | ">="
  | "<="
  | "LIKE"
  | "IN"
  | "IS";

/**
 * Supported operators for SET clause assignments (used in updates).
 */
type SetOperator = "=" | "+=" | "-=";

/**
 * Represents a single condition in a WHERE clause,
 * combining a field, a value, and an optional operator.
 *
 * @template T - The entity type extending Entity.
 */
export interface WhereCondition<T extends Entity> extends Clause<T> {
  /** The comparison operator (e.g., '=', '>', 'LIKE') */
  operator?: WhereOperator;
}

export type WhereExpression<T extends Entity> =
  | WhereCondition<T>
  | WhereComposite<T>;

/**
 * Represents a composite WHERE condition combining multiple
 * conditions with a logical connector (AND or OR).
 *
 * Example:
 * {
 *   connectors: "AND",
 *   prepositions: [
 *     { field: "userOd", operator: "=", value: "10" },
 *     { field: "age", operator: ">", value: 18 }
 *   ]
 * }
 *
 * @template T - The entity type extending Entity.
 */
export interface WhereComposite<T extends Entity> {
  connectors: "AND" | "OR";
  prepositions: WhereExpression<T>[];
}
// ...
export type Where<T extends Entity> =
  | WhereCondition<T>
  | WhereComposite<T>
  | null;
// ...
export type Join<L extends Entity, R extends Entity> = {
  /**
   * The left table being joined.
   * If this is not specified:
   * @default by DAO Entity - default it will be for the DAO that is executing the query.
   */
  leftTable?: typeof Entity;

  /**
   * The right table being joined (required).
   */
  rightTable: (new (...args: any[]) => R) & typeof Entity; // Stricter typing for class reference

  /**
   * The attributes to join on.
   * left - attribute of the left table.
   * right - attribute of the right table.
   */
  on: { left: keyof L; right: keyof R };

  /**
   * The fields of the right table being joined that you want to select.
   * These will be aliased as `${R.EntityName}_${fieldName}`.
   */
  fields?: (keyof R)[];

  /**
   * The type of join.
   */
  type?: "LEFT" | "RIGHT" | "INNER";

  /**
   * Specify where condition which involves the right hand table.
   */
  where?: JoinWhere<R>;

  /**
   * Any order depending on the joined table.
   */
  orderBy?: OrderBy<R>[];
};

/**
 * To specify the column for which the where clause is targeting during a join
 */
export type JoinWhere<T extends Entity> = {
  /**
   * The table which the where clause will target.
   * @default The right table of the Join.
   */
  targetTable?: typeof Entity;
} & Where<T>;

/**
 * Options for general queries.
 */
export type Options<T extends Entity> = {
  /**
   * Any joins for the query.
   * Note: The inner `any, any` is used to allow mixing Join types in the array.
   */
  joins?: Join<any, any>[]; // Keep `any, any` here to allow an array of mixed joins
  /**
   * Limit the results of the query
   */
  limit?: number;

  /**
   * Offset for the query
   */
  offset?: number;

  /**
   * Order the query results
   */
  orderBy?: OrderBy<T>[];
};

/**
 * Specifies an order by clause in an sql section.
 * @template T - The entity type extending Entity.
 */
export type OrderBy<T extends Entity> = {
  /**
   * The table which the attribute ordering is targeting.
   * By default this is the table which the DAO generic is for.
   */
  table?: typeof Entity;
  field: keyof T;
  direction: "ASC" | "DESC";
};

/**
 * Options specifically for selects
 */
export type SelectOptions<T extends Entity> = {
  /**
   * Specific fields to select from the main table
   */
  fields?: (keyof T)[];
} & Options<T>;